pyproject.toml
[project]
name = "finance-agentic-rag-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Chiew Jie Lun"}
]
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "ipython (>=9.5.0,<10.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-docling (>=1.0.0,<2.0.0)",
    "langchain-mcp-adapters (>=0.1.9,<0.2.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "langchain-redis (>=0.2.3,<0.3.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "markdownify (>=1.2.0,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pymupdf4llm (>=0.0.27,<0.0.28)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "redis[search] (>=4.5.0,<6.0.0)",
    "redisvl (>=0.8.1,<0.9.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
